requirements.txt
conda install -r 


request 